splits tests file
